// src/app/modelo/dto/paciente/crear-paciente-dto.ts
export class CrearPacienteDto {
    cedula: string="";
    nombre: string="";
    apellido: string="";
    telefono: string="";
    ciudad: string="";
    email: string="";
    confirmarEmail: string="";
    contrasena: string="";
    confirmarContrasena: string="";
    alergias: string="";
    foto: string="";
    tipoSangre: string="";
    eps: string="";
    fechaNacimiento: string="";
}